# storage/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: LMS App API
  description: |
    API documentation for the Library Management System mobile/web app.

    **Authentication:** Use a Bearer Token (Laravel Sanctum) for protected routes.
    Obtain a token via `POST /api/v1/auth/login`.

    **Response format:** Most endpoints return `{ message, body, ... }` where `message` is `success` or `error`.
  version: 1.0.0
  contact:
    name: LMS Backend
servers:
  - url: /api
    description: Local server
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Authentication]
      summary: Login
      description: |
        Authenticate with email or phone number and password.
        The `login` field accepts either an email address or a phone number (digits only).
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              login: ahmed@example.com
              password: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Incorrect password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              example:
                message: error
                body: كلمة المرور غير صحيحة
        '404':
          description: Email or phone not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiValidationErrorResponse'
  /v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout
      description: Revoke all tokens for the authenticated user.
      operationId: authLogout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: success
                body: تم تسجيل الخروج بنجاح
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      description: Returns the authenticated user's profile.
      operationId: authMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum Token
      description: Enter the token only (Swagger adds the Bearer prefix automatically)
...
  responses:
    Unauthorized:
      description: Unauthorized — missing or invalid token
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
              - $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          description: Email address or phone number (digits only)
          example: ahmed@example.com
        password:
          type: string
          format: password
          example: password123
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        first_name:
          type: string
          example: Ahmad
        last_name:
          type: string
          example: Mohammad
        full_name:
          type: string
          example: Ahmad Mohammad
        email:
          type: string
          format: email
          example: ahmed@example.com
        phone:
          type: string
          example: "0912345678"
        identity_number:
          type: string
          example: "1234567890"
        adress:
          type: string
          nullable: true
          example: Damascus
        role:
          type: string
          enum: [MEMBER, LIBRARIAN, ADMIN]
          example: MEMBER
        state:
          type: string
          enum: [ACTIVE, PAUSED, CANCLED]
          nullable: true
          example: ACTIVE
        photo_url:
          type: string
          format: uri
          nullable: true
          example: "http://localhost/storage/Profiles/photo.jpg"
        participe_end_date:
          type: string
          format: date
          nullable: true
          example: "2027-12-31"
        created_at:
          type: string
          format: date-time
          nullable: true
          example: "2026-06-07 10:00:00"
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          enum: [success, error]
        body:
          type: string
...
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          enum: [success]
        body:
          type: string
          example: تم تسجيل الدخول بنجاح
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          example: "1|xxxxxxxxxxxxxxxx"
    MeResponse:
      type: object
      properties:
        message:
          type: string
          enum: [success]
        body:
          type: string
        user:
          $ref: '#/components/schemas/User'
...
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
          enum: [error]
        body:
          type: string
    ApiValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
          enum: [error]
        body:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...